Use gsl::narrow_cast

+++ src/generic/activityindicator.cpp
@@ -136,7 +136,7 @@ private:
 
         // Subsequent dots are rotated by this angle with respect to the
         // previous one.
-        constexpr double angle = wxDegToRad(360. / NUM_DOTS);
+        constexpr float angle = gsl::narrow_cast<float>(wxDegToRad(360. / NUM_DOTS));
 
         // And the animation effect is achieved just by starting to draw from
         // the next position every time.
